Backend/plaid-service/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
fastapi==0.103.1
idna==3.4
nulltype==2.3.1
plaid-python==15.5.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
